refactor: Simplify expression.rs binary serialization (1244 → 719 lines)
Reduces code complexity and improves maintainability by:
- Splitting helper functions into category-based modules
- Introducing declarative macros for enum serialization
- Organizing code by functional area (operators, types, window functions)

Maintains full backward compatibility with existing serialization format.
All persistence tests pass.

Main file reduced by 42% (1244 → 719 lines).
Total code across modules: 1175 lines.

Closes #2277
